Adding instructions to create mappings
Motivation

We are currently starting to integrate mappings to other vocabularies. We
should make sure that we have a clear structure in the mapping folder and a
unified approach of handling mappings.

Modification

A markdown file was created to give instructions and examples on how to
create mappings for PaNET. Additionally, the template mapping_template.csv
has been provided.

Result

The creation of upcoming mappings has been simplified.

Closes #288