Update build procedure to create widoko documentation for releases.
Motivation:

We would like to make available documentation for all PaNET releases.

In addition, the latest release should have a permanent link, in order
to allow someone to refer to the latest release without knowing which
release that is.

Modification:

Upgrade to 0.12 of build script.  This brings the option of downloading

Enable the option of building the site documentation.

Result:

The CI/CD now provides documentation for all released versions of PaNET.